backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
fer==22.5.1
reportlab==4.0.7
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0